# optimize univ3 logic and alloc