Benchmark:
Before (after part1 https://github.com/KyberNetwork/kyberswap-dex-lib/pull/322)
```
BenchmarkCalcAmountOutV2-8                	  112780	     10659 ns/op	   11472 B/op	     180 allocs/op
BenchmarkCalcAmountOutCrossFewTickV2-8    	   67615	     17725 ns/op	   17877 B/op	     334 allocs/op
BenchmarkCalcAmountOutCrossManyTickV2-8   	   10000	    104921 ns/op	   96741 B/op	    2285 allocs/op
```

After (after this PR)
```
BenchmarkCalcAmountOutV2-8                	  240918	      4977 ns/op	    3088 B/op	      40 allocs/op
BenchmarkCalcAmountOutCrossFewTickV2-8    	  112432	     10670 ns/op	    3088 B/op	      40 allocs/op
BenchmarkCalcAmountOutCrossManyTickV2-8   	   16384	     73133 ns/op	    3089 B/op	      40 allocs/op
```

## Why did we need it?
<!--- Describe your changes in detail -->

## Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

## Release Note
<!--- Please list out special notes (if any) when releasing this commit -->
<!--- Special notes is things like breaking changes and related components, required changes and env configurations in kyber-application, etc. -->

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):